* [jsfm] refactor the normalize method
